Move KWObjectDataPath classes in specific files
De meme que KWMTDatabaseMapping, sous classe de KWDataPath, etst gere dans un fichier KWMTDatabaseMapping.h,
on deplace la classe KWObjectDataPath dans un fichier KWObjectDataPath.h:
- pour faciliter la maintenance
- probleme de dependance cyclique resolu en incluant le fichier KWClass.h dans le header
- implementation de la gestion de Get|SetDataPathManager dans KWDataPath

KWDatabase
- utilisation directe de l'objet objectDataPathManager et non d'un pointeur,
  maintenant que le probleme de dependance cyclique est resolu
